Hitmrg/develop (#38)
* fix ldap authentication

redifining the ldap authentication based on the authentication type, add ldap provider and the security configuration for application.yaml

* update securityconfig.java

i did forget an } at the end.

* ldapprovider invalid name

need to change the class name to not conflict with LdapAuthenticationProvider in compilation

* change declaration

* forced english test configuration

if tests executed in another country, tests failed so forced location to us

* authentication check

- application.yaml ldap example parameters not correctly setup
- modify AuthenticationProviderLDAP add fallback and handling errors

* Update AuthenticationProviderLDAP.java

casting DirContextOperations but Spring Security return LdapUserDetailsImpl.

So we needed to handle two types (DirContextOperations and LdapUserDetails)

* Update AuthenticationProviderLDAP.java

DefaultLdapAuthoritiesPopulator try to find AD groups in the void

* Update JwtAuthController.java

modify how the token get the username with the email if ldap authentication

* Update JwtAuthController.java

change the direction, api call error if used the previous push

* Update JwtAuthController.java

getting back to normal

* Update JwtAuthController.java

updating the token controller because if ldap authentication, the initial result with user.getEmail() parameters doest get the full email (only get the parameter)

* Update AuthenticationProviderLDAP.java

modify fetchUserAttributes to get data from ldap

* create ldap documentation

technical documentation for ldap implementation explanation and deployment with docker

* feat: Refactor LDAP configuration to separate standard Spring Boot properties from WiseMapping-specific settings

---------

Co-authored-by: hitmrg <hugo.margonelli@supinfo.com>